feat: refactor parser
