+++ packages/macros/src/attribute/with_components/mod.rs
@@ -1,0 +1,4 @@
+pub mod components;
+pub mod definition;
+pub mod diagnostics;
+pub mod parser;
